% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aquisition.R
\name{data_aquisition}
\alias{data_aquisition}
\title{Function to Acquire the data for analysis
This function imports the train file for titanic data analysis and stores them in a dataframe.}
\usage{
data_aquisition()
}
\description{
Function to Acquire the data for analysis
This function imports the train file for titanic data analysis and stores them in a dataframe.
}
