% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_graph}
\alias{plot_graph}
\title{Create plot to analyse relation function}
\usage{
Run the function by calling it without arguments.
}
\description{
This function analyses the  relationship between sex and survival.
}
\examples{
main_func()
}
