% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_data}
\alias{titanic_data}
\title{Titanic Data Set}
\format{An object of class \code{data.frame} with 891 rows and 12 columns.}
\usage{
titanic_data
}
\description{
Titanic Data Set
}
\keyword{datasets}
