% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimisition.R
\name{titanic_optimisation}
\alias{titanic_optimisation}
\title{Function to optimise the data.
This function handles the missing data, and formates the data for analysis}
\usage{
titanic_optimisation()
}
\description{
Function to optimise the data.
This function handles the missing data, and formates the data for analysis
}
