% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aquisition.R
\name{data_aquisition}
\alias{data_aquisition}
\title{Titanic Survial Rate Analysis
This package analyses the data of a list of passangers on titanic and analyses the survival rate.}
\usage{
data_aquisition()
}
\description{
Titanic Survial Rate Analysis
This package analyses the data of a list of passangers on titanic and analyses the survival rate.
}
